pyproject.toml
[project]
name = "hulo-knowledge-base"
version = "0.1.0"
description = "Hulo Language Knowledge Base with RAG + MCP + LLMs"
authors = [
    {name = "Hulo Team", email = "team@hulo-lang.org"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "markdown>=3.5.1",
    "beautifulsoup4>=4.12.2",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2",
    "tiktoken>=0.5.2",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "rich>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"] 
